Refactor ChartTheme and ChartStyles
