+++ packages/react-charts/src/components/ChartBullet/ChartBulletComparativeMeasure.tsx
@@ -5,7 +5,8 @@ import { VictoryBar } from 'victory-bar';
 import { getComparativeMeasureData } from './utils';
 import { ChartBar } from '../ChartBar';
 import { ChartContainer } from '../ChartContainer';
-import { ChartBulletStyles, ChartThemeDefinition } from '../ChartTheme';
+import { ChartThemeDefinition } from '../ChartTheme';
+import { ChartBulletStyles } from '../ChartTheme/ChartStyles';
 import { ChartTooltip } from '../ChartTooltip';
 import { getBulletComparativeMeasureTheme } from '../ChartUtils/chart-theme-types';
