+++ packages/react-charts/src/components/ChartBullet/ChartBulletTitle.tsx
@@ -3,7 +3,8 @@ import { PaddingProps, StringOrNumberOrCallback } from 'victory-core';
 import { ChartContainer } from '../ChartContainer';
 import { ChartLabel } from '../ChartLabel';
 import { ChartLegendPosition } from '../ChartLegend';
-import { ChartBulletStyles, ChartCommonStyles, ChartThemeDefinition } from '../ChartTheme';
+import { ChartThemeDefinition } from '../ChartTheme';
+import { ChartCommonStyles, ChartBulletStyles } from '../ChartTheme/ChartStyles';
 import { getBulletLabelX, getBulletLabelY } from '../ChartUtils/chart-label';
 import { getPaddingForSide } from '../ChartUtils/chart-padding';
 import { getBulletTheme } from '../ChartUtils/chart-theme-types';
