+++ packages/react-charts/src/components/ChartBullet/utils/chart-bullet-size.ts
@@ -1,5 +1,5 @@
 import { ChartThemeDefinition } from '../../ChartTheme';
-import { ChartBulletStyles } from '../../ChartTheme';
+import { ChartBulletStyles } from '../../ChartTheme/ChartStyles';
 import {
   getBulletComparativeErrorMeasureTheme,
   getBulletComparativeMeasureTheme,
