+++ packages/react-charts/src/components/ChartLegendTooltip/ChartLegendTooltipLabel.tsx
@@ -12,7 +12,7 @@ import {
   VictoryLabelProps
 } from 'victory-core';
 import { ChartLabel } from '../ChartLabel';
-import { ChartLegendTooltipStyles } from '../ChartTheme';
+import { ChartLegendTooltipStyles } from '../ChartTheme/ChartStyles';
 
 /**
  * ChartLegendLabel renders a legend tooltip label
