+++ packages/react-charts/src/components/ChartTheme/ChartTheme.ts
@@ -1,7 +1,7 @@
 import { VictoryThemeDefinition } from 'victory-core';
 
 // Note: Victory incorrectly typed ThemeBaseProps.padding as number instead of PaddingProps
-export interface ChartThemeDefinitionInterface extends VictoryThemeDefinition {}
+interface ChartThemeDefinitionInterface extends VictoryThemeDefinition {}
 
 /**
  * Chart theme definition
