+++ packages/react-charts/src/components/ChartTheme/index.ts
@@ -1,4 +1,2 @@
-export * from './ChartTheme';
-export * from './ChartThemeColor';
-export * from './ChartThemeTypes';
-export * from './ChartStyles';
+export type { ChartThemeDefinition } from './ChartTheme';
+export { ChartThemeColor } from './ChartThemeColor';
